requirements.txt
discord.py
asyncpg
fastapi[all]
python-dotenv
emoji
